+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [4.5.0]
+### Added
+- Config option for invite email delay. Default is no delay.
+- Purge inactive users after a configurable amount of time.
+
+### Changed
+- Renamed UserResponse property `groups` to `member` to coincide with SAML2 naming.
+
 ## [4.4.0] - 2019-05-29
 ### Added
 - Use `supportName` and `supportEmail` if `helpCenterUrl` is not provided